Updated scripts for provider_updates
