--- data_updates/pe_providers.psql
@@ -1,1 +1,0 @@
-SELECT npi from provider_effort.providers;
